services/product-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiosqlite==0.19.0

pytest==7.4.3
pytest-asyncio==0.21.1
